requirements.txt
selenium
pytest
pytest-bdd
pyyaml
pytest-html